requirements.txt
alembic==1.0.11
aniso8601==7.0.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
bcrypt==3.1.7
cffi==1.12.3
Click==7.0
coverage==4.3
Flask==1.1.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
importlib-metadata==0.19
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
lazy-object-proxy==1.4.1
Mako==1.0.14
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-sqlalchemy==0.17.0
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.1
pluggy==0.12.0
psycopg2-binary==2.8.3
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
pylint-plugin-utils==0.5
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.0.1
pytest-cov==2.7.1
pytest-flask==0.15.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.6
typed-ast==1.4.0
wcwidth==0.1.7
Werkzeug==0.15.5
wrapt==1.11.2
zipp==0.5.2
